refactored views into its own module
