+++ portal/views/__init__.py
@@ -1,0 +1,4 @@
+from .form import *
+from .categories import *
+from .application import *
+from .public import *
